+++ lang_qc/db/helper/wells.py
@@ -23,11 +23,14 @@ import logging
 from datetime import date, datetime, timedelta
 from typing import ClassVar, List
 
-from pydantic import BaseModel, Extra, Field
+from pydantic import BaseModel, ConfigDict, Field
 from sqlalchemy import and_, or_, select
 from sqlalchemy.orm import Session
 
-from lang_qc.db.helper.qc import get_qc_states_by_id_product_list
+from lang_qc.db.helper.qc import (
+    get_qc_states_by_id_product_list,
+    qc_state_for_product_exists,
+)
 from lang_qc.db.mlwh_schema import PacBioRunWellMetrics
 from lang_qc.db.qc_schema import QcState, QcStateDict, QcType
 from lang_qc.models.pacbio.well import PacBioPagedWells, PacBioWell
@@ -42,6 +45,8 @@ This package is using an undocumented feature of Pydantic, type
 Here this type is used to mark a purely internal to the class variables.
 """
 
+INBOX_LOOK_BACK_NUM_WEEKS = 12
+
 
 class WellWh(BaseModel):
     """
@@ -54,12 +59,9 @@ class WellWh(BaseModel):
         title="SQLAlchemy Session",
         description="A SQLAlchemy Session for the ml warehouse database",
     )
-    INBOX_LOOK_BACK_NUM_WEEKS: ClassVar = 4
-
-    class Config:
-        allow_mutation = False
-        arbitrary_types_allowed = True
-        allow_population_by_field_name = True
+    model_config = ConfigDict(arbitrary_types_allowed=True, populate_by_name=True)
+    # frozen=True from Pydantic v2 does not work the way we want it to during testing.
+    # The TestClient seems to be keeping these instances alive and changing them.
 
     def get_mlwh_well_by_product_id(
         self, id_product: str
@@ -77,31 +79,35 @@ class WellWh(BaseModel):
 
     def recent_completed_wells(self) -> List[PacBioRunWellMetrics]:
         """
-        Get recent completed wells from the mlwh database.
-        The implementation of the inbox query might change when the QC outcomes
-        become available in mlwh.
+        Get recent not QC-ed completed wells from the mlwh database.
+        Recent wells are defined as wells that completed within the
+        last 12 weeks.
         """
 
         ######
         # It is important not to show aborted wells in the inbox.
         #
-        # The well can be complete as in Illumina 'run complete' but that's not
-        # the same as analysis complete which the other conditions are trying for.
-        # It potentially gets a bit easier with v11 but those conditions should
-        # still work ok.
+        # The well can be complete, but that's not the same as analysis
+        # complete which the other conditions are trying for.
+        # It potentially gets a bit easier with v11 but those conditions
+        # should still work ok.
         #
 
         # Using current local time.
         # Generating a date rather than a timestamp here in order to have a consistent
         # earliest date for the look-back period during the QC team's working day.
-        my_date = date.today() - timedelta(weeks=self.INBOX_LOOK_BACK_NUM_WEEKS)
+        my_date = date.today() - timedelta(weeks=INBOX_LOOK_BACK_NUM_WEEKS)
         look_back_min_date = datetime(my_date.year, my_date.month, my_date.day)
+
+        # Select the wells that has not been QC-ed, but later double-check against
+        # the LangQC database.
 
         # TODO: fall back to run_complete when well_complete is undefined
 
         query = (
             select(PacBioRunWellMetrics)
             .where(PacBioRunWellMetrics.well_status == "Complete")
+            .where(PacBioRunWellMetrics.qc_seq_state.is_(None))
             .where(PacBioRunWellMetrics.run_complete > look_back_min_date)
             .where(PacBioRunWellMetrics.polymerase_num_reads.is_not(None))
             .where(
@@ -179,10 +185,7 @@ class PacBioPagedWellsFactory(WellWh, PagedResponse):
         ),
     }
 
-    class Config:
-        arbitrary_types_allowed = True
-        extra = Extra.forbid
-        allow_mutation = True
+    model_config = ConfigDict(arbitrary_types_allowed=True, extra="forbid")
 
     def create_for_qc_status(
         self, qc_flow_status: QcFlowStatusEnum
@@ -211,6 +214,8 @@ class PacBioPagedWellsFactory(WellWh, PagedResponse):
             QcFlowStatusEnum.UNKNOWN,
         ]:
             wells = self._aborted_and_unknown_wells(qc_flow_status)
+        elif qc_flow_status == QcFlowStatusEnum.UPCOMING:
+            wells = self._upcoming_wells()
         else:
             wells = self._get_wells_for_status(qc_flow_status)
 
@@ -320,6 +325,60 @@ class PacBioPagedWellsFactory(WellWh, PagedResponse):
             else:
                 well.copy_run_tracking_info(db_well)
 
+    def _upcoming_wells(self):
+        """
+        Upcoming wells are recent wells, which do not belong to any other
+        QC flow statuses as defined in QcFlowStatus. Recent wells are defined
+        as wells that belong to runs that started within the last 12 weeks.
+        """
+
+        recent_completed_product_ids = [
+            w.id_pac_bio_product for w in self.recent_completed_wells()
+        ]
+
+        my_date = date.today() - timedelta(weeks=INBOX_LOOK_BACK_NUM_WEEKS)
+        look_back_min_date = datetime(my_date.year, my_date.month, my_date.day)
+
+        # If queries for any other filters change, this query should be revised
+        # since we are repeating (but negating) a few condition that are
+        # associated with some of the statuses (filters).
+
+        query = (
+            select(PacBioRunWellMetrics)
+            .where(PacBioRunWellMetrics.run_start > look_back_min_date)
+            .where(PacBioRunWellMetrics.qc_seq_state.is_(None))
+            .where(
+                PacBioRunWellMetrics.id_pac_bio_product.not_in(
+                    recent_completed_product_ids
+                )
+            )
+            .where(PacBioRunWellMetrics.well_status.not_like("Abort%"))
+            .where(PacBioRunWellMetrics.well_status.not_like("Terminat%"))
+            .where(PacBioRunWellMetrics.well_status.not_like("Fail%"))
+            .where(PacBioRunWellMetrics.well_status.not_like("Error%"))
+            .where(PacBioRunWellMetrics.well_status.not_in(["Unknown", "On hold"]))
+            .order_by(
+                PacBioRunWellMetrics.run_start,
+                PacBioRunWellMetrics.pac_bio_run_name,
+                PacBioRunWellMetrics.plate_number,
+                PacBioRunWellMetrics.well_label,
+            )
+        )
+
+        wells = []
+        for w in self.session.execute(query).scalars().all():
+            if (
+                qc_state_for_product_exists(
+                    session=self.qcdb_session, id_product=w.id_pac_bio_product
+                )
+                is False
+            ):
+                wells.append(w)
+
+        self.total_number_of_items = len(wells)  # Save the number of retrieved wells.
+
+        return self._well_models(self.slice_data(wells), False)
+
     def _recent_inbox_wells(self, recent_wells):
 
         inbox_wells_indexes = []
@@ -397,7 +456,7 @@ class PacBioPagedWellsFactory(WellWh, PagedResponse):
                 # A well can have only one or zero current sequencing outcomes.
                 if qced_products is not None:
                     attrs["qc_state"] = qced_products[0]
-            pb_well = PacBioWell.parse_obj(attrs)
+            pb_well = PacBioWell.model_validate(attrs)
             pb_well.copy_run_tracking_info(db_well)
             pb_wells.append(pb_well)